GUVNOR-2552: Guided Decision Table Editor: Fix column Synchronizers
