+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/main/java/org/drools/workbench/screens/guided/dtable/client/widget/table/model/converters/column/impl/DescriptionColumnConverter.java
@@ -42,7 +42,7 @@ public class DescriptionColumnConverter extends BaseColumnConverterImpl {
         final GridColumn<?> uiColumn = newStringColumn( makeHeaderMetaData( column ),
                                                         Math.max( column.getWidth(),
                                                                   DEFAULT_COLUMN_WIDTH ),
-                                                        false,
+                                                        true,
                                                         true,
                                                         access,
                                                         gridWidget );
