+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/main/java/org/drools/workbench/screens/guided/dtable/client/widget/table/model/synchronizers/impl/ActionSetFieldColumnSynchronizer.java
@@ -35,6 +35,11 @@ public class ActionSetFieldColumnSynchronizer extends ActionColumnSynchronizer {
     }
 
     @Override
+    public boolean handlesAppend( final MetaData metaData ) throws ModelSynchronizer.MoveColumnVetoException {
+        return handlesUpdate( metaData );
+    }
+
+    @Override
     public boolean handlesUpdate( final MetaData metaData ) throws ModelSynchronizer.MoveColumnVetoException {
         if ( !( metaData instanceof ColumnMetaData ) ) {
             return false;
