+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/main/java/org/drools/workbench/screens/guided/dtable/client/widget/table/model/synchronizers/impl/ActionWorkItemColumnSynchronizer.java
@@ -34,6 +34,11 @@ public class ActionWorkItemColumnSynchronizer extends ActionColumnSynchronizer {
     }
 
     @Override
+    public boolean handlesAppend( final MetaData metaData ) throws ModelSynchronizer.MoveColumnVetoException {
+        return handlesUpdate( metaData );
+    }
+
+    @Override
     public boolean handlesUpdate( final MetaData metaData ) throws ModelSynchronizer.MoveColumnVetoException {
         if ( !( metaData instanceof ColumnMetaData ) ) {
             return false;
