+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/main/java/org/drools/workbench/screens/guided/dtable/client/widget/table/model/synchronizers/impl/BRLActionColumnSynchronizer.java
@@ -34,15 +34,12 @@ public class BRLActionColumnSynchronizer extends BaseColumnSynchronizer<BaseColu
 
     @Override
     public int priority() {
-        return 1;
+        return 3;
     }
 
     @Override
     public boolean handlesAppend( final MetaData metaData ) {
-        if ( !( metaData instanceof ColumnMetaData ) ) {
-            return false;
-        }
-        return ( (ColumnMetaData) metaData ).getColumn() instanceof BRLActionColumn;
+        return handlesUpdate( metaData );
     }
 
     @Override
