+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/main/java/org/drools/workbench/screens/guided/dtable/client/widget/table/model/synchronizers/impl/BRLConditionColumnSynchronizer.java
@@ -35,15 +35,12 @@ public class BRLConditionColumnSynchronizer extends BaseColumnSynchronizer<BaseC
 
     @Override
     public int priority() {
-        return 1;
+        return 3;
     }
 
     @Override
     public boolean handlesAppend( final MetaData metaData ) {
-        if ( !( metaData instanceof ColumnMetaData ) ) {
-            return false;
-        }
-        return ( (ColumnMetaData) metaData ).getColumn() instanceof BRLConditionColumn;
+        return handlesUpdate( metaData );
     }
 
     @Override
