+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/main/java/org/drools/workbench/screens/guided/dtable/client/widget/table/model/synchronizers/impl/ConditionColumnSynchronizer.java
@@ -60,7 +60,7 @@ public class ConditionColumnSynchronizer extends BaseColumnSynchronizer<PatternC
 
     @Override
     public boolean handlesAppend( final MetaData metaData ) {
-        return metaData instanceof PatternConditionMetaData;
+        return handlesUpdate( metaData );
     }
 
     @Override
