+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/main/java/org/drools/workbench/screens/guided/dtable/client/widget/table/model/synchronizers/impl/MetaDataColumnSynchronizer.java
@@ -32,10 +32,7 @@ public class MetaDataColumnSynchronizer extends BaseColumnSynchronizer<BaseColum
 
     @Override
     public boolean handlesAppend( final MetaData metaData ) {
-        if ( !( metaData instanceof ColumnMetaData ) ) {
-            return false;
-        }
-        return ( (ColumnMetaData) metaData ).getColumn() instanceof MetadataCol52;
+        return handlesUpdate( metaData );
     }
 
     @Override
